# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Life House Reentry Finance Portal API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /admin/users:
    get:
      operationId: listAdminUsers
      tags: [admin]
      summary: List all user accounts (admin only)
      responses:
        "200":
          description: Users
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUserListResponse"
    post:
      operationId: createAdminUser
      tags: [admin]
      summary: Create a new user account (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAdminUserBody"
      responses:
        "201":
          description: Created user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUserResponse"
  /admin/users/{id}/password:
    post:
      operationId: changeAdminUserPassword
      tags: [admin]
      summary: Change a user account password (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChangeAdminUserPasswordBody"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkResponse"
...
  /admin/restore-day:
    post:
      operationId: restoreToday
      tags: [admin]
      summary: Restore database to today's start-of-day snapshot (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MasterPasswordBody"
      responses:
        "200":
          description: Restore result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestoreDayResponse"
  /admin/wipe-data:
    post:
      operationId: wipeAllData
      tags: [admin]
      summary: Permanently delete all transactional data (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MasterPasswordBody"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkResponse"
  /admin/email-settings:
    get:
      operationId: getEmailSettings
      tags: [admin]
      summary: Get the configurable email templates and sender name (admin only)
      responses:
        "200":
          description: Email settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmailSettingsResponse"
    put:
      operationId: updateEmailSettings
      tags: [admin]
      summary: Replace the email templates and sender name (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmailSettingsUpdateBody"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkResponse"
  /admin/email-settings/test:
    post:
      operationId: sendEmailSettingsTest
      tags: [admin]
      summary: Send a test email of one template to the current admin user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmailSettingsTestBody"
      responses:
        "200":
          description: Test email send result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmailSettingsTestResponse"
...
  /admin/notifications/{id}/resend:
    post:
      operationId: resendAdminNotification
      tags: [admin]
      summary: Re-send the email for a specific notification (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Resent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResendAdminNotificationResponse"
        "502":
          description: Email could not be sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResendAdminNotificationResponse"
...
components:
  schemas:
...
    AuthUser:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        role:
          type: string
          # `service` is the automation account authenticated via
          # INTEGRATION_API_KEY bearer token. It is hidden from the
          # admin user list and never returned by /auth/me or /auth/login,
          # but is included here so the type union stays consistent
          # with the server.
          enum: [admin, approver, submitter, service]
      required:
        - id
        - email
        - firstName
        - lastName
        - role
    AuthUserResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/AuthUser"
      required: [user]
...
    AdminUserListResponse:
      type: object
      properties:
        users:
          type: array
          items:
            $ref: "#/components/schemas/AuthUser"
      required: [users]
    CreateAdminUserBody:
      type: object
      properties:
        email:
          type: string
          format: email
        firstName:
          type: string
          minLength: 1
        lastName:
          type: string
          minLength: 1
        role:
          type: string
          enum: [admin, approver, submitter]
        password:
          type: string
          minLength: 6
      required: [email, firstName, lastName, role, password]
    ChangeAdminUserPasswordBody:
      type: object
      properties:
        password:
          type: string
          minLength: 6
      required: [password]
...
    OkResponse:
      type: object
      properties:
        ok:
          type: boolean
      required: [ok]
    MasterPasswordBody:
      type: object
      properties:
        masterPassword:
          type: string
      required: [masterPassword]
...
    RestoreDayResponse:
      type: object
      additionalProperties: true
      properties:
        ok:
          type: boolean
      required: [ok]
    EmailTemplate:
      type: object
      properties:
        type:
          type: string
        subject:
          type: string
        body:
          type: string
        defaultSubject:
          type: string
        defaultBody:
          type: string
        variables:
          type: array
          items:
            type: string
        sampleVariables:
          type: object
          additionalProperties:
            type: string
        updatedAt:
          type: string
          nullable: true
      required:
        - type
        - subject
        - body
        - defaultSubject
        - defaultBody
        - variables
        - sampleVariables
    EmailSettingsResponse:
      type: object
      properties:
        senderName:
          type: string
        defaultSenderName:
          type: string
        templates:
          type: array
          items:
            $ref: "#/components/schemas/EmailTemplate"
      required: [senderName, defaultSenderName, templates]
    EmailTemplateUpdate:
      type: object
      properties:
        type:
          type: string
        subject:
          type: string
        body:
          type: string
      required: [type, subject, body]
    EmailSettingsUpdateBody:
      type: object
      properties:
        senderName:
          type: string
        templates:
          type: array
          items:
            $ref: "#/components/schemas/EmailTemplateUpdate"
      required: [senderName, templates]
    EmailSettingsTestBody:
      type: object
      properties:
        type:
          type: string
        subject:
          type: string
        body:
          type: string
      required: [type, subject, body]
    EmailSettingsTestResponse:
      type: object
      properties:
        ok:
          type: boolean
        delivered:
          type: boolean
        to:
          type: string
        subject:
          type: string
        body:
          type: string
        note:
          type: string
      required: [ok, delivered]
...
    ResendAdminNotificationResponse:
      type: object
      properties:
        ok:
          type: boolean
        status:
          type: string
        error:
          type: string
      required: [status]
...